cleaning up!!!
got a workable app :)
