--- file_output/test.csv
@@ -1,69 +1,0 @@
-
-1671461928.654579,2022-12-19 09:58:48,yy
-1671462006.702069,2022-12-19 10:00:06,yy
-1671462007.513551,2022-12-19 10:00:07,
-1671462015.946024,2022-12-19 10:00:15,testing my file writer
-1671462033.2510378,2022-12-19 10:00:33,
-1671462264.994122,2022-12-19 10:04:24,uuu
-1671462410.7265768,2022-12-19 10:06:50,ooo
-1671462445.979054,2022-12-19 10:07:25,uuu
-1671463819.450815,2022-12-19 10:30:19,kk
-1671463890.527584,2022-12-19 10:31:30,ii
-1671463895.758698,2022-12-19 10:31:35,iii
-1671463957.809223,2022-12-19 10:32:37,yuup
-1671468076.3811998,2022-12-19 11:41:16,test
-1671468239.1121888,2022-12-19 11:43:59,test
-1671468368.4200299,2022-12-19 11:46:08,i
-1671468383.3116539,2022-12-19 11:46:23,ii
-1671468433.928246,2022-12-19 11:47:13,test
-1671468519.8058681,2022-12-19 11:48:39,Text
-1671468545.7552798,2022-12-19 11:49:05,I'm going to take a much longer note, one that tkes up multiple lines and see what it looks like
-1671468583.806061,2022-12-19 11:49:43,text input
-1671468589.5371509,2022-12-19 11:49:49,text input
-1671468662.536484,2022-12-19 11:51:02,smaller note here
-1671468674.98601,2022-12-19 11:51:14,let's 
-1671468675.898597,2022-12-19 11:51:15,keep
-1671468676.639573,2022-12-19 11:51:16,going
-1671468677.548042,2022-12-19 11:51:17,and
-1671468678.446887,2022-12-19 11:51:18,see
-1671468679.353473,2022-12-19 11:51:19,what
-1671468680.849859,2022-12-19 11:51:20,happens
-1671468787.513967,2022-12-19 11:53:07,1
-1671468787.980428,2022-12-19 11:53:07,2
-1671468788.390625,2022-12-19 11:53:08,3
-1671468788.762997,2022-12-19 11:53:08,4
-1671468789.072865,2022-12-19 11:53:09,5
-1671468789.4648058,2022-12-19 11:53:09,1
-1671468789.840491,2022-12-19 11:53:09,2
-1671468790.239813,2022-12-19 11:53:10,
-1671468790.773524,2022-12-19 11:53:10,5
-1671468791.478663,2022-12-19 11:53:11,
-1671468791.659916,2022-12-19 11:53:11,
-1671468791.808747,2022-12-19 11:53:11,
-1671468809.527935,2022-12-19 11:53:29,
-1671468809.699828,2022-12-19 11:53:29,
-1671468809.8498838,2022-12-19 11:53:29,
-1671468810.010556,2022-12-19 11:53:30,
-1671468810.166514,2022-12-19 11:53:30,
-1671468810.3159978,2022-12-19 11:53:30,
-1671468810.474358,2022-12-19 11:53:30,
-1671468810.6456509,2022-12-19 11:53:30,
-1671468810.8234031,2022-12-19 11:53:30,
-1671468811.152761,2022-12-19 11:53:31,
-1671468920.242047,2022-12-19 11:55:20,Note taker here. Input text. Happy fun happy happy fun fun 
-1671468953.285352,2022-12-19 11:55:53,1
-1671468953.630568,2022-12-19 11:55:53,
-1671468953.788269,2022-12-19 11:55:53,
-1671468953.945769,2022-12-19 11:55:53,
-1671468954.089889,2022-12-19 11:55:54,
-1671468954.2212968,2022-12-19 11:55:54,
-1671468954.380373,2022-12-19 11:55:54,
-1671468954.534677,2022-12-19 11:55:54,
-1671468954.736803,2022-12-19 11:55:54,
-1671469269.6738482,2022-12-19 12:01:09,yes
-1671469628.598989,2022-12-19 12:07:08,test
-1671469631.758525,2022-12-19 12:07:11,test
-1671469909.042191,2022-12-19 12:11:49,
-1671469909.645514,2022-12-19 12:11:49,
-1671470229.0063841,2022-12-19 12:17:09,test
-1671470681.980768,2022-12-19 12:24:41,notes !
