--- moderated_usability_note_taker_TUI/notetaker.css
@@ -1,13 +1,0 @@
-Screen {
-    align: center middle;
-}
-
-Input {
-    width: 40;
-    height: 9;
-    padding: 1 2;
-    background: $panel;
-    color: $text;
-    border: $secondary tall;
-    content-align: center middle;
-}
