--- stopwatch_TUI/stopwatch.css
@@ -1,53 +1,0 @@
-Stopwatch {
-    layout: horizontal;
-    background: $boost;
-    height: 5;
-    min-width: 50;
-    margin: 1;
-    padding: 1;
-}
-
-TimeDisplay {
-    content-align: center middle;
-    text-opacity: 60%;
-    height: 3;
-}
-
-Button {
-    width: 16;
-}
-
-#start {
-    dock: left;
-}
-
-#stop {
-    dock: left;
-    display: none;
-}
-
-#reset {
-    dock: right;
-}
-
-.started {
-    text-style: bold;
-    background: $success;
-    color: $text;
-}
-
-.started TimeDisplay {
-    text-opacity: 100%;
-}
-
-.started #start {
-    display: none
-}
-
-.started #stop {
-    display: block
-}
-
-.started #reset {
-    visibility: hidden
-}
