--- stopwatch_TUI/stopwatch02.css
@@ -1,1 +1,0 @@
-/* Blank for now */
