--- stopwatch_TUI/stopwatch03.css
@@ -1,31 +1,0 @@
-Stopwatch {
-    layout: horizontal;
-    background: $boost;
-    height: 5;
-    margin: 1;
-    min-width: 50;
-    padding: 1;
-}
-
-TimeDisplay {
-    content-align: center middle;
-    text-opacity: 60%;
-    height: 3;
-}
-
-Button {
-    width: 16;
-}
-
-#start {
-    dock: left;
-}
-
-#stop {
-    dock: left;
-    display: none;
-}
-
-#reset {
-    dock: right;
-}
