--- textual-tutorial/static_update.css
@@ -1,3 +1,0 @@
-TextDisplay {
-    border: wide green;
-}
